MAINT: move utils to trading_calendars.utils
